MA-11423[Android] refine power HAL to support all big-Little cores on imx8qm
The iMX8QM is big-Little core MPU and the A53 cores and A72 cores
have separate cpufreq governor to control the frequency.
cpu4/cpu5 is A72 core on imx8qm, cpu0 is the default core.
The governor of A72 core will keep the same as A53 core's.

When this HAL initialize, it find the governor of each cluster according to
the policy in /sys/devices/system/cpu/cpufreq. Each policy has separate
governor for its cpus.

Change-Id: I03aa341b5ca781715733c01cfb2e373a0883ee5e
Signed-off-by: Zhang Bo <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/4160
Reviewed-by: zhang sanshan <sanshan.zhang@nxp.com>