MA-18674-1 Enable runtime cpu hotplug in Thermal hal
remove cpu hotplug implementation in kernel space, change it to thermal HAL.
add the implementation in sendThermalChangedCallback, it is originally a callback
registered for watching throttling, will loop to update temp, we can set runtime
cpu hotplug here.

if current temp is higher than tempThreshold.hotThrottlingThresholds[3], will
offline related CPUs, this is defined in thermal json file, and when temp throttle
and lower than the last trip point tempThreshold.hotThrottlingThresholds[2],
will online the defined CPUs. and this is align with the implementation in imx-cpufreq-dt.ko.

Change-Id: I7d13e534cf1a4f86640b0275307de8808f67de36
Signed-off-by: Dandan Sun <dandan.sun@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/15085
Reviewed-by: zhang sanshan <sanshan.zhang@nxp.com>