MA-18798-2 thermal/sensor: Fix Unchecked return value from library
this patch is to fix [Coverity CID :11903805/25570] "Unchecked return
value from library(CHECKED_RETURN)".

it failed for "Calling fcntl(this->uevent_fd_.operator int(), 4, 2048)
without checking return value. This library function may fail and
return an error code. and If this is not checked, the error condition
may not be handled correctly."

now add the condition check and handle the error condition.

Change-Id: Ic00678d919ac3386b8bfe79620b2a4fa960190f3
Signed-off-by: Dandan Sun <dandan.sun@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/15268
Reviewed-by: zhang sanshan <sanshan.zhang@nxp.com>