MA-18798-5 sensor: Fix Array compared against 0 issue
this patch is to fix [Coverity CID :25557/25558] "Array
compared against 0(NO_EFFECT)".

it failed for "The address of an array is never NULL,
so the comparison will always evaluate the same way."

in FSLSensorsHub.h, mClassPath is defined as char
mClassPath[sensors][PATH_MAX]; in previous implemention, compare
mClassPath[0] with NULL, because mClassPath[0] return the
address of the array. so the coverity check failed.

to fix this issue, compare the mClassPath[what][0] with '\0'.

Change-Id: I308461611e7073a4bd6c7ce7730a920de6bdec45
Signed-off-by: Dandan Sun <dandan.sun@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/15271
Reviewed-by: zhang sanshan <sanshan.zhang@nxp.com>