MA-18745[Android_8MQ_EVK]Display: UI flicker when palyback two video in split screen mode
When playback two video in split mode(one in Gallery, one in Cactus player),
the UI will flicker sometimes. It's because there are two surfaceview layer
be set true for isOverlay variable. So one of it will not be set as DEVICE
comsition type, while other UI layer set as DEVICE type for performance when
UI is not change. It cause framework will setRenderTarget only with that CLIENT
video frame, and no UI information( because UI layer set as DEVICE type).
When UI updated, all layers will composite to framebuffer. So the flicker
issue happen.
Here clear the isOverlay variable when one framebuffer commit.

Change-Id: I9de31c68715c09bfa983b408d9556ba855d56591
Signed-off-by: Zhang Bo <bo.zhang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/15383
Reviewed-by: juan.hao <juan.hao@nxp.com>