MA-19002 Use fence in camera HAL
If acquire_fence is valid, wait on it before the buffer is available to feed.
For release_fence, since the buffer is ready before return to framwork,
currently no need to use it.

Since the fence need be duped in SubmitRequests(), add struct "FrameRequest" to
record both HwlPipelineRequest and related fences.

Change-Id: I982ffb3ae15c1ad75cd60b2a83e55e0bd5b4968f
Signed-off-by: Fang Hui <hui.fang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/15729
Reviewed-by: zhang sanshan <sanshan.zhang@nxp.com>
Reviewed-by: guoyin.chen <guoyin.chen@nxp.com>