MA-19095 eliminate logs when switch usb functions
Check whether the udc device exists when bind configfs with udc devices.
This can eliminate the log below:
    udc-core: couldn't find an available UDC or it's busy

Set the poll timeout value based on whether configfs is already binded
with the udc device, if it is not, set it to be 2000, which means 2s, it
is used in the condition for boards whose usb driver makes the udc
device disappear if the board is disconnected from the host. because
later there maybe no events if adbd is not enabled. If configfs is
already binded to the udc device, set the timeout value to be -1.

Delay the operation of unlinking usb functions until previous monitorFfs
thread is killed, so below logs can be eliminated:
    Config b/1 of g1 needs at least one function.

Change-Id: Ic303d2e159b9be8b7092a22779df2cd17ca2cc05
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/15999
Reviewed-by: Tian Yang <yang.tian@nxp.com>