MA-19095-1 use dummy udc to eliminate logs
use dummy udc to eliminate logs when the board is disconnected from the
host.

When the function to be set includes adbd and the udc device set in
property "vendor.usb.config" does not exist, which may happen on some
boards when the boards are disconnected from the host, try to bind
configfs with dummy udc, then adbd can receive "FUNCTIONFS_BIND" event
and it won't be in a loop to configure adbd functionfs again and again
so logs can be eliminated.

also add watch on the usb controller sysfs path, so when the board is
connected to the host, usb gadget HAL can bind the configfs with the
valid udc again.

Change-Id: I6508ab4ff16eb19e5e14bf67ed519b299aa561c6
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/16015
Reviewed-by: Tian Yang <yang.tian@nxp.com>