MA-19309 Raw data picture is black
On 8mp, due to 60fps performance, we change HW engine from G3D to G2D.
For G2D copy, if the out buffer is read by CPU, seems the cached is not
flushed, so the picture is black.
Also, to correct the color revert issue for RAW format, use Revert16BitEndian()
in handleFrameByGPU_2D(), just as handleFrameByGPU_3D().

Change-Id: I09b532b2588049c91ad201448da90f9a7b355970
Signed-off-by: Fang Hui <hui.fang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/16749
Reviewed-by: zhang sanshan <sanshan.zhang@nxp.com>