MA-19605 Fix a issue occured when RNDIS is set twice
When setting the usb gadget configuration to RNDIS with below steps, the
RNDIS function is set in continuous two times. The second one would fail
then and usb gadget does not work.

  Settings > System > Developer options > Default USB configuration

This is a fix to that issue, when create the RNDIS function link, if the
link exist, do not treat it as a failure.

Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>
Change-Id: I4214c8d4ae8e912ccca424a63a082b68208c4b37
Reviewed-on: http://androidsource.ap.freescale.net/project/17428
Reviewed-by: Wang Haoran <elven.wang@nxp.com>